+++ pkg/crypto/tls_test.go
@@ -9,11 +9,11 @@ import (
 	"github.com/aws/eks-anywhere/pkg/crypto"
 )
 
-var validCipherSuitesString = "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256"
+var validCipherSuites = []string{"TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256"}
 
 func TestSecureCipherSuiteNames(t *testing.T) {
-	string := crypto.SecureCipherSuitesString()
-	if !reflect.DeepEqual(string, validCipherSuitesString) {
-		assert.Equal(t, validCipherSuitesString, string, "cipher suites don't match")
+	list := crypto.SecureCipherSuiteNames()
+	if !reflect.DeepEqual(list, validCipherSuites) {
+		assert.Equal(t, validCipherSuites, list, "expected cipher suites to be list of valid")
 	}
 }