+++ docs/content/en/docs/reference/changelog.md
@@ -15,9 +15,8 @@ menu:
 
 ### Changed
 
-- Kubernetes components and etcd now use TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256 as the
-  configured TLS cipher suite [#657](https://github.com/aws/eks-anywhere/pull/657), 
-  [#759](https://github.com/aws/eks-anywhere/pull/759)
+- Kubernetes components and local etcd now use TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256 as the
+  configured TLS cipher suite [#657](https://github.com/aws/eks-anywhere/pull/657)
 
 ## v0.6.0 - 2021-10-29